+++ R/Map_Invalsi.R
@@ -80,7 +80,8 @@ Map_Invalsi <- function(data = NULL, Year = 2023, subj_toplot = "ITA", grade = 8
   }
 
   while(is.null(data)){
-    data <- Get_Invalsi_IS(level = level, verbose = verbose, autoAbort = autoAbort)
+    data <- Get_Invalsi_IS(level = level, verbose = verbose,
+                           multiple_out = FALSE, autoAbort = autoAbort)
     if(is.null(data)){
       if(!autoAbort){
         holdOn <- ""
@@ -95,6 +96,18 @@ Map_Invalsi <- function(data = NULL, Year = 2023, subj_toplot = "ITA", grade = 8
           cat("You chose to retry \n")
         }
       } else return(NULL)
+    }
+  }
+
+  if(!is.data.frame(data) && is.list(data)){
+    if (toupper(level) %in% c("LAU", "NUTS-4", "MUNICIPALITY", "MUN")){
+      data <- data$Municipality_data
+    } else{
+      data <- data$Province_data
+    }
+    if(is.null(data)){
+      message("No data provided at the ", level, " level")
+      return(NULL)
     }
   }